refactor(testmessage): make testmessage into module
